feat: scroll item to the edit text
